bet_maker/requirements.txt
aerich==0.7.2
aio-pika==9.4.2
aiohttp==3.9.5
aiohttp-retry==2.8.3
aiomysql==0.2.0
aiormq==6.8.0
aiosignal==1.3.1
aiosmtplib==3.0.1
aiosqlite==0.17.0
annotated-types==0.6.0
anyio==4.3.0
async-timeout==4.0.3
asyncmy==0.2.9
attrs==23.2.0
bcrypt==4.1.2
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==42.0.5
dictdiffer==0.9.0
dnspython==2.6.1
email_validator==2.1.1
exceptiongroup==1.2.0
fastapi==0.110.1
fastapi-utils==0.2.1
frozenlist==1.4.1
greenlet==3.0.3
h11==0.14.0
idna==3.7
iso8601==1.1.0
multidict==6.0.5
pamqp==3.3.0
pycparser==2.22
pydantic==1.10.15
pydantic_core==2.18.1
PyJWT==2.8.0
PyMySQL==1.1.0
pyotp==2.9.0
pypika-tortoise==0.1.6
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.3
sniffio==1.3.1
SQLAlchemy==1.4.52
starlette==0.37.2
tomlkit==0.12.4
tortoise-orm==0.20.0
twilio==9.2.2
typing_extensions==4.11.0
urllib3==2.2.2
uvicorn==0.29.0
yarl==1.9.4
